WIP:Frontend UI working
