+++ client/components/itemtable/itemTable.js
@@ -10,6 +10,7 @@ import { DeletePackItemModal } from '../pack_table/DeletePackItemModal';
 import { PaginationLimit } from '../paginationChooseLimit';
 import Loader from '../Loader';
 import useCustomStyles from '~/hooks/useCustomStyles';
+import {loadStyles} from './itemsTable.style'
 
 export const ItemsTable = ({
   limit,
@@ -112,7 +113,7 @@ export const ItemsTable = ({
           {isLoading ? (
             <Loader />
           ) : (
-            data.items.map((item, index) => {
+            data.map((item, index) => {
               return <TableItem key={index} itemData={item} />;
             })
           )}
@@ -157,71 +158,3 @@ export const ItemsTable = ({
   );
 };
 
-const loadStyles = (theme) => {
-  const { currentTheme } = theme;
-  return {
-    container: {
-      flex: 1,
-      padding: 10,
-      width: '100%',
-    },
-    tableStyle: {
-      width: '100%',
-      paddingHorizontal: 20,
-    },
-    mainTitle: {
-      marginTop: 10,
-      marginBottom: 10,
-      alignItems: 'center',
-      justifyContent: 'center',
-    },
-    categoryRow: {
-      padding: 10,
-      borderRadius: 5,
-      alignItems: 'center',
-      justifyContent: 'flex-start',
-    },
-    title: {
-      height: 50,
-      backgroundColor: currentTheme.colors.primary,
-      borderRadius: 10,
-      justifyContent: 'center',
-      paddingLeft: 15,
-    },
-    titleText: {
-      fontWeight: 'bold',
-      color: currentTheme.colors.text,
-    },
-    head: {
-      height: 50,
-      borderBottomWidth: 1,
-      borderBottomColor: '#D1D5DB',
-      borderTopLeftRadius: 10,
-      borderTopRightRadius: 10,
-    },
-    headerText: {
-      fontWeight: 'bold',
-      color: '#000000',
-    },
-    row: {
-      flexDirection: 'row',
-      height: 60,
-      alignItems: 'center',
-      backgroundColor: '#FFFFFF',
-      borderBottomWidth: 1,
-      borderBottomColor: '#D1D5DB',
-    },
-    infoContainer: {
-      flexDirection: 'row',
-      justifyContent: 'space-between',
-      padding: 25,
-      backgroundColor: '#F8F8F8',
-    },
-    noItemsText: {
-      fontWeight: 'bold',
-      fontSize: 16,
-      marginTop: 20,
-      textAlign: 'center',
-    },
-  };
-};
